split operators for propositional logic and ldlf formulas.
All the elementary operator like Not, And, Or, Implies
and Equivalence for the LDLf parser have been splitted in two.

In particular, if the operator is applied to a propositional formula,
then the strings to be used are (as usual):
- not: !
- and: &
- or: |
- implies: ->
- equivalence: <->

whereas, for LDLf formulas:
- not: !!
- and: &&
- or: ||
- implies: -->
- equivalence: <-->

that is, double the character of the operator.
This change is due to introduce the following abbreviation:

given P a propositional formula:

P -> <P>tt

However, this introduced several conflicts, mostly due to our use
of PLY. This is going to change in the near future.